refactor: renderJoke abstraction
add renderJoke abstraction for test saftey

References #FSA22V1-74

On branch feat--onLoad-renderJoke
Changes to be committed:
modified:   src/js/index.js
modified:   src/js/index.spec.js
new file:   src/js/renderJoke.js
